Moved CompareTimingModels to a different file AGAIN
